requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cloudinary==1.31.0
Django==4.1.5
django-cors-headers==3.13.0
djangorestframework==3.14.0
factory-boy==3.2.1
Faker==16.6.1
Pillow==9.4.0
psycopg2==2.9.5
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.14
